debugging

+++ schemas/+meso/SegmentationChunks.m
@@ -1,7 +1,7 @@
 %{
 # registration between different segmentation chunks within a recording
 -> meso.Segmentation
-segmentation_chunk_id  : int    # id for the subsection of the recording this segmentation is for, for cases with multi-chunk segemntation (e.g. because of z drift)
+segmentation_chunk_id  : tinyint    # id for the subsection of the recording this segmentation is for, for cases with multi-chunk segemntation (e.g. because of z drift)
 ---
 tif_file_list          : blob   # cell array with names of tif files that went into this chunk
 imaging_frame_range    : blob   # [firstFrame lastFrame] of this chunk with respect to the full session
